refactor: organize imports and remove unused code
Clean up the codebase by removing unused imports, dead code, and redundant comments.

Changes:
- **Parser**:
  - Removed unused imports (`Token`, `TokenType`) from multiple parser modules where they were not needed.
  - Deleted unused `parse_asm_value` helper function.
  - Cleaned up commented-out code and Korean comments in `decl.rs`, `functions.rs`, and `io.rs`.
- **Lexer**:
  - Removed redundant `pub use common::*;` in `core.rs` and `scan.rs`.
- **Codegen**:
  - Removed unused `AddressSpace` import in `statement/variable.rs`.
  - Deleted panic branches for unreachable code in `wave_type_to_llvm_type`.
  - Removed Korean comments in `statement/io.rs`.
- **General**:
  - Bumped version to `0.1.6-pre-beta`.
  - Standardized import grouping across modules.

This maintenance commit improves code quality and reduces compiler warnings.
