+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wavec"
-version = "0.1.5-pre-beta"
+version = "0.1.6-pre-beta"
 edition = "2021"
 
 [lib]
