+++ front/lexer/src/lexer/core.rs
@@ -1,5 +1,4 @@
 use crate::token::TokenType;
-use super::common::*;
 
 #[derive(Debug, Clone)]
 pub struct Token {
@@ -39,7 +38,7 @@ impl<'a> Lexer<'a> {
     pub fn tokenize(&mut self) -> Vec<Token> {
         let mut tokens = Vec::new();
         loop {
-            let token = self.next_token(); // scan.rs에 구현
+            let token = self.next_token();
             if token.token_type == TokenType::Eof {
                 tokens.push(token);
                 break;
