+++ front/parser/src/lib.rs
@@ -3,7 +3,7 @@ pub mod ast;
 pub mod format;
 pub mod import;
 pub mod stdlib;
-pub mod type_system;
 pub mod verification;
+pub mod expr;
 
 pub use parser::*;
